app/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-logging     == 2.2.0
google-api-python-client == 1.12.8
google-auth              == 1.23.0
google-auth-httplib2     == 0.0.4
netaddr                  == 0.8.0
requests                 == 2.24.0